Rename to media, complete docstrings, change rhoCp to SI (fromkSI)
