+++ doc/source/media.rst
@@ -1,0 +1,10 @@
+.. media:
+
+************
+Media Module
+************
+
+.. automodule:: pygfunction.media
+    :members:
+    :undoc-members:
+    :show-inheritance:
