--- doc/source/properties.rst
@@ -1,10 +1,0 @@
-.. properties:
-
-*******************
-Properties Module
-*******************
-
-.. automodule:: pygfunction.properties
-    :members:
-    :undoc-members:
-    :show-inheritance:
