+++ pygfunction/__init__.py
@@ -4,7 +4,7 @@ from . import boreholes
 from . import gfunction
 from . import heat_transfer
 from . import load_aggregation
+from . import media
 from . import networks
 from . import pipes
 from . import utilities
-from . import properties
